backend/requirements.txt
fastapi
uvicorn[standard]
watchfiles
python-multipart
google-genai
qdrant-client
PyMuPDF
python-dotenv
pydantic
